move popgym to its own file
