+++ examples/02_popgym_suite.py
@@ -3,7 +3,7 @@ from argparse import ArgumentParser
 import wandb
 
 import amago
-from amago.envs.builtin.gym_envs import POPGymEnv
+from amago.envs.builtin.popgym_envs import POPGymEnv
 from amago.cli_utils import *
